+++ user/src/stat.c
@@ -7,13 +7,13 @@ int mkdir(const char *path, mode_t mode) {
     register unsigned long x8 asm("x8") = 136;
     asm volatile("svc #0"
         : "+r"(x0) : "r"(x1), "r"(x1) : "memory");
-    return (int)x0;
+    return (int)(long)x0;
 }
 
 int stat(const char *path, struct stat *buf) {
     register const char *x0 asm("x0") = path;
     register struct stat *x1 asm("x1") = buf;
-    register uint64_t x8 asm("x8") = 38;  // Your syscall number
+    register uint64_t x8 asm("x8") = 38;
     asm volatile("svc #0"
         : "+r"(x0) : "r"(x1), "r"(x8) : "memory");
     return (int)(long)x0;
